[20] Move all children functions to s/children.go
